21-01-2024

+++ database/migrations/2023_12_16_080551_create_barang_mentahs_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('barangmentahs', function (Blueprint $table) {
             $table->id();
             $table->string('nama');
-            $table->integer('jumlah');
+            $table->double('jumlah');
             $table->string('satuan');
             $table->integer('harga');
             $table->timestamps();
