+++ database/migrations/2023_12_16_081702_create_barangjadis_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama');
             $table->string('ukuran')->nullable();
-            $table->integer('jumlah');
+            $table->double('jumlah');
             $table->integer('harga');
             $table->timestamps();
         });
